tests/tools/mock-agent/pyproject.toml
[project]
name = "mock-coding-agent"
version = "0.1.0"
description = "A mock coding agent that emits thinking traces, tool-use messages, and writes Codex-compatible session files; includes a mock OpenAI/Anthropic API server."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Zahary Karadjov", email = "zahary@blocksense.network" }]
dependencies = [
    "pyyaml>=6.0",
    "yamllint>=1.32.0",
]

[project.optional-dependencies]
test = ["pytest", "pexpect"]

[project.scripts]
mockagent = "src.cli:main"
